Refactor SSD
Summary:
Adjust the network architecture to more closely match that used in the tensorflow example. (Note: This breaks weights saved from previous versions of the architecture.) As part of this, adjust the format of y_true: it now has two parts, and the network has two outputs (a location part and a categorization part). This means adjustments to the SSD utilities and the SSD loss function as well.

Also makes the hyperparameters more closely match those from the paper, including adding an option for weight decay.

Adds test for SSD loss and roundtripping SSD bounding box data.

Reviewers: jeremy

Reviewed By: jeremy

Subscribers: jenkins

Differential Revision: https://phb.1026labs.com/D1145