Add Min and Assign aggregation ops
Summary:
Add < (min) and = (assign) aggregation operations to those allowed in contractions. Switch Keras backend to use < and = contractions where appropriate. Make Assign assert that each output entry is assigned to at most once; add a test of this.

Note that the Tile compiler is not very smart in determining multiple assignment for =. Use + or no_defract as appropriate to bypass this.

Fix the spelling of gradient.

Reviewers: jeremy

Reviewed By: jeremy

Differential Revision: https://phb.1026labs.com/D1159