Set the right target to be used after cpu got split into llvm_cpu and opencl_cpu
Use llvm_cpu when running NNs.
The target/deviceID got changed from "cpu" to "llvm_cpu" and "opencl_cpu", but the platform config script didn't change.
This results in the following error when trying to run resnet50 NN.
  [libprotobuf FATAL external/com_google_protobuf/src/google/protobuf/map.h:1060] CHECK failed: it != end(): key not found: cpu
  CHECK failed: it != end(): key not found: cpu

This change sets the target/device used for the NN run to be llvm_cpu.