Add an extra prune_idxs pass to remove unused indexes (#691)
* Fix tile-contract configuration for CPU Jit

Make the AutoTilePass configuration to split properly the blocks, so the multotreading is done on non accumulating blocks. This ay there are no race conditions.

Fixed useage of LLVM voidPtrType since it is asserting in the master branch of LLVM.

* Add an extra prune_idxs pass to remove unused indexes

After auto-tiling add a pass to remove unused indexes. This shaves down the resnet50 infer time by 12-13%.