Add prng signness test ex (#1053)
* Add a PRNG signness test.

* Fixed a potential type issue.

* Fix Linux optimizer issue.

This issue happens only in optimized Linux builds - the dbg ones are fine.
The issue is an optimization that the C++ compiler makes on Linux that generates the wrong code. It has to do with the StridedMemRef<XXX, 1> in the signature of the RT PRNG function.
In this particular case, the compiler uses the value of the type parameter to optimize the loop for calculating the count of outputs and doesn't fill all the values.
A bigger value (like 20) should be sufficient enough and after talking to @tzerrell we agree on that.

* Fixed minor issues.

Co-authored-by: Lubomir Litchev <Lubomir.Litchev@intel.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>