Add AffineVectoreReduce instruction to the pxa layer. (#1231)
* Add AffineVectoreReduce instruction to the pxa layer.

This is needed to allow us to use the data flow analysis mechanisms that the AffineReduceOp exposes/utilizes, so Fusion and other passes can operate on the AffineVectorReduceOp.

* Added more tests.

Removed the NoSideEffects, because the Red/Write semantics of the instruction have bigger weight then that flag.

* Templatize a struct and take out some code.

* Set the map attribute on the vector_load/store instructions.

Make the tests to check more things.

* Fixed typos.

* Stop recapturing and compare the captured values in  the tests.