[pxa] Add separate GPU thread function
This change adds separate function for threading parallel op for GPU -
tiling into global and local loops.
Such function is useful when new operation needs to be added after
normal threading pass, such as when inserting data reorders.
It then needs to be manually threaded to comply with expected semantics.