WIP: JSON config for multithreading
This patch is the first of many to support
Tiling and parallelization in PlaidML from
an external description. This patch adds
support for multithreading by specifying
which parallel loop can run on separate
OpenMP threads.

1. The patch identifies parallel loops
uniquely by associating hierarchical indices
to them. For example, if there are two affine
parallel loops in the IR, they are assigned
indices 0 and 1. Further, if the first loop
has a nested affine parallel loop, it gets the id
"0,0".

2. A JSON file containing loop ids and their
schedules which need to be parallelized is
committed.