TPSS: parallelization directives
This patch adds support for parallelization directives to be
specified in a file using the environment <PLAIDML_PARALLELIZATION_CONFIG_FILE>.
This patch adds a rule parser which matches the shapes of convolution
based on the equalities/inequalities in the config file and applies the rules
that follow. Parallelization rule is partially supported as of now in this patch.
Subsequent changes with include collapse/serialize directives' support.