Allow users to define CC/CXX options for CMake on configure
Users that don't have a supported compiler (gcc 9.4) as a system
compiler can still pass the --cc and --cxx (or use standard environment
variables CC and CXX) to set it for the PlaidML build without having to
change their paths.

This, and the feedback on #1972, make it build successfully and pass all
the tests.

Fixes #1972